+++ services/ui-src/src/components/index.tsx
@@ -2,7 +2,7 @@ export { App } from "./app/App";
 export { AppRoutes } from "./app/AppRoutes";
 export { Banner } from "./banners/Banner";
 export { DueDateTable } from "./tables/DueDateTable";
-export { DueDateDropdown } from "./accordion/DueDateDropdown";
+export { DueDateDropdown } from "./accordions/DueDateDropdown";
 export { Footer } from "./layout/Footer";
 export { Header } from "./layout/Header";
 export { Icon } from "./icons/Icon";
@@ -11,4 +11,4 @@ export { LoginIDM } from "./logins/LoginIDM";
 export { Menu } from "./menus/Menu";
 export { MenuOption } from "./menus/MenuOption";
 export { RouterLink } from "./app/RouterLink";
-export { TemplateCard } from "./templateCard/TemplateCard";
+export { TemplateCard } from "./templateCards/TemplateCard";
