+++ services/ui-src/src/components/templateCards/data/MCPARCardText.json
