+++ services/ui-src/src/views/Home/Home.tsx
@@ -6,6 +6,7 @@ import { Banner, TemplateCard } from "../../components/index";
 // utils
 import { UserRoles } from "utils/types/types";
 import { useUser } from "utils/auth";
+import { bannerTypes } from "utils/types/types";
 
 export default () => {
   const { userRole, userState } = useUser();
@@ -20,7 +21,7 @@ export default () => {
       {userState ? (
         <Box sx={sx.root} data-testid="home-view">
           <Banner
-            // status="info"
+            status={bannerTypes.INFO}
             bgColor="palette.alt_lightest"
             accentColor="palette.alt"
             title="Welcome to the new Managed Care Reporting tool!"
